Remove no longer needed class LeagueGenerator
All logic inside has been moved to the harness test
