+++ src/League/Tests/LeagueHarnessTest.php
@@ -1,38 +1,60 @@
 <?php
 
+    use League\Match;
+    use League\Team;
+    use League\League;
+    use League\Round;
+
     class LeagueHarnessTest extends \PHPUnit_Framework_TestCase {
+
+        private $league;
 
         protected function setUp() {
             $this->setFixedRandomSeed();
+            $teams = [
+                new Team('The Creative Prowlers'),
+                new Team('The Brute Comets'),
+                new Team('The Heavenly Warhawks'),
+                new Team('Spirit Squirrels'),
+                new Team('Courageous Donkeys'),
+                new Team('Magic Zebras'),
+                new Team('Magic Hurricanes'),
+                new Team('Storm Hedgehogs'),
+                new Team('Young Monarchs'),
+                new Team('Kalonian Hydras'),
+                new Team('Eager Beluga Whales'),
+                new Team('Ancient Trolls')
+            ];
+
+            $this->league = new League($teams);
+
+            $this->league->populateRounds();
         }
 
         /**
          * @test
          */
         public function should_generate_66_total_matches_for_12_teams() {
-            $leagueGenerator = new \League\LeagueGenerator();
-            $generation = $leagueGenerator->generate();
-            $matches = $generation['matches'];
-            $this->assertEquals(66, $matches);
+            $matchesCount = array_reduce($this->league->rounds, [$this,'sumAllMatches'], 0);
+
+            $this->assertEquals(66, $matchesCount);
         }
 
         /**
          * @test
          */
         public function should_generate_11_rounds_for_12_teams() {
-            $leagueGenerator = new \League\LeagueGenerator();
-            $generation = $leagueGenerator->generate();
-            $rounds = $generation['rounds'];
-            $this->assertEquals(11, $rounds);
+            $roundsCount = count($this->league->rounds);
+            $this->assertEquals(11, $roundsCount);
         }
 
         /**
          * @test
          */
         public function should_generate_6_matches_per_round_for_12_teams() {
-            $leagueGenerator = new \League\LeagueGenerator();
-            $generation = $leagueGenerator->generate();
-            $matchesPerRound = $generation['matchesPerRound'];
+            $roundsCount = count($this->league->rounds);
+            $matchesCount = array_reduce($this->league->rounds, [$this,'sumAllMatches'], 0);
+            $matchesPerRound = $matchesCount / $roundsCount;
             $this->assertEquals(6, $matchesPerRound);
         }
 
@@ -40,15 +62,12 @@
          * @test
          */
         public function first_match_should_be_Young_Monarchs_Against_Storm_Hedgehogs() {
-            $leagueGenerator = new \League\LeagueGenerator();
-            $generation = $leagueGenerator->generate();
-            $league = $generation['league'];
-            $firstMatch = $league->rounds[0]->matches[0];
+            $firstMatch = $this->league->rounds[0]->matches[0];
 
-            $youngMonarchs = new \League\Team("Young Monarchs");
-            $stromHedgehogs = new \League\Team("Storm Hedgehogs");
+            $youngMonarchs = new Team("Young Monarchs");
+            $stromHedgehogs = new Team("Storm Hedgehogs");
 
-            $expectedFirstMatch = new \League\Match($youngMonarchs, $stromHedgehogs, 2, 0);
+            $expectedFirstMatch = new Match($youngMonarchs, $stromHedgehogs, 2, 0);
             $this->assertEquals($expectedFirstMatch, $firstMatch);
         }
 
@@ -57,21 +76,22 @@
          * @test
          */
         public function last_match_home_team_should_be_The_Heavenly_Warhawks_Against_The_Brute_Comets() {
-            $leagueGenerator = new \League\LeagueGenerator();
-            $generation = $leagueGenerator->generate();
-            $league = $generation['league'];
-            $lastRound = array_pop($league->rounds);
+            $lastRound = array_pop($this->league->rounds);
             $lastMatch = array_pop($lastRound->matches);
 
-            $theHeavenlyWarhawks = new \League\Team("The Heavenly Warhawks");
-            $theBruteComets = new \League\Team("The Brute Comets");
+            $theHeavenlyWarhawks = new Team("The Heavenly Warhawks");
+            $theBruteComets = new Team("The Brute Comets");
 
-            $expectedLastMatch = new \League\Match($theHeavenlyWarhawks, $theBruteComets, 1, 0);
+            $expectedLastMatch = new Match($theHeavenlyWarhawks, $theBruteComets, 1, 0);
             $this->assertEquals($expectedLastMatch, $lastMatch);
         }
 
         private function setFixedRandomSeed() {
             mt_srand(0);
 
+        }
+
+        private function sumAllMatches(int $totalMatchesCount, Round $currentRound){
+            return $totalMatchesCount + count($currentRound->matches);
         }
     }
